semantic-search-api-durable-functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-core
openai==0.27.2
azure-kusto-data==4.0.2
pandas
pyodbc
azure-search-documents==11.4.0b6
azure-ai-formrecognizer
python-dotenv
faiss-cpu
tiktoken
azure-storage-blob
tenacity